refactor: 로그인/로그아웃 버튼 컴퍼넌트화 #9
