+++ tsconfig.app.json
@@ -1,4 +1,6 @@
 {
+  "extends": "./tsconfig.json",
+
   "compilerOptions": {
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "target": "ES2022",
